KS-2-5 Задание готово для проверки №1

+++ src/main/kotlin/lesson_2/task_5.kt
@@ -8,12 +8,12 @@ import kotlin.math.round
 fun main() {
     val contribution: Int = 70_000
     val interestRate: Float = 16.7f
-    val years: Int = 20
+    val years: Byte = 20
 
 
     val dd = floor((1f + interestRate / 100.0) * 1000.0) / 1000.0 // получение double без лишних значений
 
-    val amount = contribution * dd.pow(years)
+    val amount = contribution * dd.pow(years.toInt())
 
     println("Размер вклада через $years лет составит ${String.format("%.3f", amount)} рублей.")
 }
